+++ custom_components/yale_lock_manager/const.py
@@ -2,7 +2,7 @@
 from typing import Final
 
 DOMAIN: Final = "yale_lock_manager"
-VERSION: Final = "1.2.1.0"
+VERSION: Final = "1.3.0.0"
 
 # Z-Wave JS Domain
 ZWAVE_JS_DOMAIN: Final = "zwave_js"